% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMCMC.R
\name{ConstructMCMCArray}
\alias{ConstructMCMCArray}
\title{Read in JAGS objects}
\usage{
ConstructMCMCArray(
  ChainIDs,
  n.samplestot.max = 15000,
  core.run = FALSE,
  use.informative.priors = FALSE
)
}
\arguments{
\item{ChainIDs}{Optional: specify which chains to include
(to use when you want to exclude a chain that crashed, or which has not finished yet).}

\item{n.samplestot.max}{Maximum number of posterior samples to save}

\item{core.run}{Set to TRUE if running core data}

\item{output.dir}{Directory where MCMC output was stored and will be stored.}
}
\value{
Invisible
}
\description{
Read in JAGS objects and constructs \code{mcmc.array}, which is saved to \code{output}.
}
\examples{
ConstructMCMCArray(ChainIDs = seq(1,3))
}
