% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_diagnostics.R
\name{eff_size}
\alias{eff_size}
\title{Run MCMC diagnostics - effective sample size}
\usage{
eff_size(mcmc.array, pars.check)
}
\arguments{
\item{mcmc.array}{an mcmc array from \code{\link{run_core}}}

\item{pars.check}{the parameter names to be checked}
}
\value{
A message will appear about convergence
}
\description{
Run MCMC diagnostics - effective sample size
}
