% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_diagnostics.R
\name{gr_diag}
\alias{gr_diag}
\title{Run MCMC diagnostics - rhat}
\usage{
gr_diag(mcmc.array, pars.check)
}
\arguments{
\item{mcmc.array}{an mcmc array from \code{\link{run_core}}}

\item{pars.check}{the parameter names to be checked}
}
\value{
A message will appear about convergence
}
\description{
Run MCMC diagnostics - rhat
}
