% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_diagnostics.R
\name{mcse}
\alias{mcse}
\title{Run MCMC diagnostics - monte carlo standard error (mcse)}
\usage{
mcse(mcmc.array, pars.check)
}
\arguments{
\item{mcmc.array}{an mcmc array from \code{\link{run_core}}}

\item{pars.check}{the parameter names to be checked}
}
\value{
A message will appear about convergence
}
\description{
Run MCMC diagnostics - monte carlo standard error (mcse)
}
