% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_core_species}
\alias{NJ_core_species}
\title{Core species counts from New Jersey, USA.}
\format{
A data frame with 69 rows and 24 variables:
\describe{
  \item{Ab}{species count}
  \item{Ad}{species count}
  \item{Ai}{species count}
  \item{Am}{species count}
  \item{As}{species count}
  \item{Calc}{species count}
  \item{Ea}{species count}
  \item{Hs}{species count}
  \item{Jm+Bp}{species count}
  \item{Mf}{species count}
  \item{Mp}{species count}
  \item{PH}{species count}
  \item{Ph}{species count}
  \item{Pi}{species count}
  \item{Pl}{species count}
  \item{Rs}{species count}
  \item{TG}{species count}
  \item{TO}{species count}
  \item{Tc}{species count}
  \item{Ti+Sl}{species count}
  \item{Tq}{species count}
  \item{Ts}{species count}
  \item{Tt}{species count}
  \item{Depth}{core deth, cm}
}
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_core_species
}
\description{
A dataset containing the raw core counts of foraminifera species for field study sites in New Jersey, USA
}
\keyword{datasets}
