% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_modern_elevation}
\alias{NJ_modern_elevation}
\title{Modern elevations from New Jersey, USA.}
\format{
A data frame with 175 rows and 3 variables:
\describe{
  \item{Site}{Name of field site}
  \item{SampleID}{ID for the sample}
  \item{SWLI}{Standardised water level index (standardised elevation)}

}
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_modern_elevation
}
\description{
A dataset containing the elevations corresponding to the modern assemblages for field study sites in New Jersey, USA
}
\keyword{datasets}
