% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_modern_species}
\alias{NJ_modern_species}
\title{Modern species counts from New Jersey, USA.}
\format{
A data frame with 175 rows and 23 variables:
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_modern_species
}
\description{
A dataset containing the raw modern counts of foraminifera species for field study sites in New Jersey, USA
}
\keyword{datasets}
