% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_priors}
\alias{NJ_priors}
\title{Prior elevation ranges for core assemblages}
\format{
A data frame with 69 rows and 3 variables:
\describe{
  \item{Depth}{core depth, cm}
  \item{prior_lwr}{elevation lower bound}
  \item{prior_upr}{elevarion upper bound}

}
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_priors
}
\description{
A dataset containing upper and lower bounds for core elevations based on a secondary proxy for field study sites in New Jersey, USA
}
\keyword{datasets}
