% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_responsecurves.R
\name{response_curves}
\alias{response_curves}
\title{Create Species Response Curves}
\usage{
response_curves(modern_mod, species_select = NULL)
}
\arguments{
\item{modern_mod}{An object of class \code{BTF} from \code{\link{run_modern}}}

\item{species_select}{a vector of species names for which you want to create response curves}
}
\value{
Response curve data files (empirical and model) and SRC plots
}
\description{
Create Species Response Curves
}
\examples{
test_modern_mod <- run_modern(modern_elevation = NJ_modern_elevation,
                         modern_species = NJ_modern_species,
                          n.iter = 10,
                          n.burnin = 1,
                          n.thin = 1)
response_curves(test_modern_mod)
}
