% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_valid.R
\name{run_valid}
\alias{run_valid}
\title{Carry out a validation}
\usage{
run_valid(
  modern_elevation = NULL,
  modern_species = NULL,
  n_folds = 1,
  use_uniform_prior = FALSE,
  dx = 0.2
)
}
\arguments{
\item{modern_elevation}{A dataframe of modern elevations}

\item{modern_species}{A dataframe of modern counts (to be sorted with \code{\link{sort_modern}})}

\item{n_folds}{Number of folds for CV}

\item{use_uniform_prior}{change prior on elevation to be uniform}

\item{dx}{The elevation interval for spacing the spline knots. Defaults to 0.2.}
}
\value{
A table of validation results
}
\description{
Carry out a validation
}
