% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_sort_data.R
\name{sort_core}
\alias{sort_core}
\title{Sort core dataset to be compatible with the sorted modern data}
\usage{
sort_core(core_dat, species_names)
}
\arguments{
\item{coredata}{A data object containing core species counts}

\item{speciesnames}{A list of species names that have been returned from \code{\link{sortmoderndata}}}
}
\value{
Returns the sorted coredata where the species columns are compatible with the modern data that has been returned from \code{\link{sortmoderndata}}
}
\description{
Sort core dataset to be compatible with the sorted modern data
}
\examples{
corecounts_sorted<-sortcoredata(coredata=core_counts,speciesnames)
}
