% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_sort_data.R
\name{sort_modern}
\alias{sort_modern}
\title{Sort modern species dataset based on total count size}
\usage{
sort_modern(modern_dat)
}
\arguments{
\item{modern_dat}{A modern data object containing species counts}
}
\value{
Returns a modern dataset that has been sorted based on species count size, as well as the sorted species names and the index for the first species that contains all zero counts (if any)
}
\description{
Sort modern species dataset based on total count size
}
