% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_swliresults.R
\name{swli_results}
\alias{swli_results}
\title{Get SWLI estimates and uncertainty}
\usage{
swli_results(obj)
}
\arguments{
\item{obj}{An object of class \code{BTF} from \code{\link{run_core}}}
}
\value{
Table of results and plot
}
\description{
Get SWLI estimates and uncertainty
}
\examples{
test_modern_mod <- run_modern(modern_elevation = NJ_modern_elevation,
                                modern_species = NJ_modern_species,
                                n.iter = 10,
                                n.burnin = 1,
                                n.thin = 1)
test_core_mod <- run_core(test_modern_mod,
                          core_species = NJ_core_species,
                          n.iter = 10,
                          n.burnin = 1,
                          n.thin = 1)
swli_results(test_core_mod)
}
