% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_core_species}
\alias{NJ_core_species}
\title{Core species counts from New Jersey, USA.}
\format{
A data frame with 69 rows and 24 variables:
\describe{
  \item{Depth}{core deth, cm}
}
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_core_species
}
\description{
A dataset containing the raw core counts of foraminifera species for field study sites in New Jersey, USA
}
\keyword{datasets}
