% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_responsecurves.R
\name{response_curves}
\alias{response_curves}
\title{Create Species Response Curves}
\usage{
response_curves(obj, species_select = NULL)
}
\arguments{
\item{obj}{An object of class \code{BTF} from \code{\link{run_modern}}}

\item{species_select}{a vector of species names for which you want to create response curves}
}
\value{
Response curve data files (empirical and model) and SRC plots
}
\description{
Create Species Response Curves
}
\examples{
responsecurves()
}
