% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_runcore.R
\name{run_core}
\alias{run_core}
\title{Run the core model}
\usage{
run_core(
  obj,
  core_counts = NULL,
  prior_el = NULL,
  dx = 0.2,
  ChainNums = seq(1, 3),
  n.iter = 15000,
  n.burnin = 1000,
  n.thin = 7,
  validation.run = FALSE,
  use_uniform_prior = FALSE
)
}
\arguments{
\item{obj}{An object of class \code{BTF} from \code{\link{run_modern}}}

\item{dx}{Parameter for spacing of spline knots}

\item{ChainNums}{The number of MCMC chains}

\item{n.iter}{The number of MCMC iterations}

\item{n.burnin}{The number of burnin MCMC samples}

\item{n.thin}{The number of thinning}

\item{core_counts.csv}{A .csv file containing core species counts}

\item{run.on.server}{Set to TRUE for running chains in parallel on a server}

\item{use.informative.priors}{Set to TRUE if prior information is available}
}
\value{
Output will be saved to \code{output.dir}
}
\description{
Run the core model
}
\examples{
coremodel<-runcore(core_counts.csv = "core_counts.csv")
}
