% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_run_modern.R
\name{run_modern}
\alias{run_modern}
\title{Run the modern (calibration) model}
\usage{
run_modern(
  modern_elevation.csv = NULL,
  modern_counts.csv = NULL,
  dx = 0.2,
  ChainNums = seq(1, 3),
  n.iter = 70000,
  n.burnin = 20000,
  n.thin = 30,
  run.on.server = FALSE,
  validation.run = FALSE,
  fold = 1
)
}
\arguments{
\item{modern_elevation.csv}{A .csv file location for modern elevations}

\item{modern_counts.csv}{A .csv file location for modern counts (to be sorted with \code{\link{sort_modern}})}

\item{dx}{The elevation interval for spacing the spline knots. Defaults to 0.2.}

\item{ChainNums}{The number of MCMC chains to run}

\item{n.iter}{The number of iterations}

\item{n.burnin}{The number of burnin samples}

\item{n.thin}{The number of thinning}

\item{run.on.server}{Set to TRUE if you wish to run the chains in parallel on a server}
}
\value{
Nothing is returned, the relevant data will be saved.
}
\description{
Run the modern (calibration) model
}
\examples{
run_modern(modern_elevation.csv = 'elevation.csv',
modern_counts.csv = 'counts.csv',
n.iter=10,
n.thin = 1,
n.burnin = 1)
}
