% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_valid.R
\name{run_valid}
\alias{run_valid}
\title{Carry out a 10-fold cross validation}
\usage{
run_valid(modern_elevation, modern_counts)
}
\arguments{
\item{modern_elevation}{}

\item{modern_counts}{}
}
\value{
A table of validation results
}
\description{
Carry out a 10-fold cross validation
}
\examples{
run_valid()
}
