% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_valid.R
\name{run_valid}
\alias{run_valid}
\title{Carry out a validation}
\usage{
run_valid(
  modern_elevation = NULL,
  modern_species = NULL,
  n_folds = 1,
  use_uniform_prior = FALSE,
  dx = 0.2,
  modern.iter = 40000,
  modern.burnin = 10000,
  modern.thin = 15,
  core.iter = 15000,
  core.burnin = 1000,
  core.thin = 7,
  scale_x = FALSE
)
}
\arguments{
\item{modern_elevation}{A dataframe of modern elevations}

\item{modern_species}{A dataframe of modern counts (to be sorted with \code{\link{sort_modern}})}

\item{n_folds}{Number of folds for CV}

\item{use_uniform_prior}{change prior on elevation to be uniform}

\item{dx}{The elevation interval for spacing the spline knots. Defaults to 0.2.}

\item{modern.iter}{The number of iterations for the modern model}

\item{modern.burnin}{The number of burnin samples for the modern model}

\item{modern.thin}{The number of thinning for the modern model}

\item{core.iter}{The number of iterations for the core model}

\item{core.burnin}{The number of burnin samples for the core model}

\item{core.thin}{The number of thinning for the core model}

\item{scale_x}{Set to TRUE to scale elevation data to have mean 0 and sd 1}
}
\value{
A table of validation results
}
\description{
Carry out a validation
}
