% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_valid.R
\name{run_valid}
\alias{run_valid}
\title{Carry out a validation}
\usage{
run_valid(
  modern_elevation = NULL,
  modern_species = NULL,
  n_folds = 1,
  use_uniform_prior = FALSE
)
}
\arguments{
\item{modern_elevation}{}

\item{modern_species}{}

\item{n_folds}{}
}
\value{
A table of validation results
}
\description{
Carry out a validation
}
\examples{
run_valid()
}
