% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_sort_data.R
\name{sort_core}
\alias{sort_core}
\title{Sort core dataset to be compatible with the sorted modern data}
\usage{
sort_core(core_dat, species_names)
}
\arguments{
\item{core_dat}{A data object containing core species counts}

\item{species_names}{A list of species names that have been returned from \code{\link{sort_modern}}}
}
\value{
Returns the sorted core data where the species columns are compatible with the modern data that has been returned from \code{\link{sort_modern}}
}
\description{
Sort core dataset to be compatible with the sorted modern data
}
