% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_swliresults.R
\name{swli_results}
\alias{swli_results}
\title{Get SWLI estimates and uncertainty}
\usage{
swli_results(obj)
}
\arguments{
\item{obj}{An object of class \code{BTF} from \code{\link{run_core}}}
}
\value{
Table of results and plot
}
\description{
Get SWLI estimates and uncertainty
}
\examples{
swliresults()
}
